refactor(shared): Split event files into domain-based modules for LLM optimization
Split 7 large event files (400-800+ lines each) into smaller, focused modules:

- combat.ts (819 lines) → combat/initiation, phases, blocking, attacks, resolution
- sites.ts (771 lines) → sites/adventure, rewards, interaction, special
- lifecycle.ts (742 lines) → lifecycle/game, round, turn, time
- progression.ts (578 lines) → progression/fame, levels, skills
- tactics.ts (571 lines) → tactics/selection, activation, rest
- units.ts (525 lines) → units/recruitment, activation, health
- mana.ts (441 lines) → mana/dice, tokens, crystals

Each original file now re-exports from its subdirectory for backward compatibility.
This follows the same pattern as previous refactorings (validators, validActions, commands).
